git-svn-id: https://www.ekylibre.org/svn/trunk/ekylibre@703 67a09383-3dfa-4221-8551-890bac9c277c

+++ app/models/stock_location.rb
@@ -65,9 +65,11 @@ class StockLocation < ActiveRecord::Base
       if !product_stock[0].nil?
         reception = (self.product_id == product_id || product_stock[0].current_real_quantity <= 0)
         self.update_attributes!(:product_id=>product_id) if product_stock[0].current_real_quantity <= 0
+        #if product_stock[0].current_real_quantity <= 0
         for ps in product_stock
-          ps.destroy
+          ps.destroy if ps.product_id != product_id and ps.current_real_quantity <=0
         end
+        #end
       else
         self.update_attributes!(:product_id=>product_id)
       end
